% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2.R
\name{current}
\alias{current}
\title{Current Weather}
\usage{
current(x)
}
\arguments{
\item{x}{City name, country code}
}
\value{
City name, date, weather, temperature, pressure, humidity, visibility, wind speed
}
\description{
Current Weather
}
\examples{
current("vancouver")
current("vancouver, ca")
current("Vancouver, CA")
current("VANCOUVER, CA")
}
