% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w1.R
\name{fiveDay}
\alias{fiveDay}
\title{5 days weather forecast}
\usage{
fiveDay(x)
}
\arguments{
\item{x}{City name, country code}
}
\value{
City name, country code, date, weather description, temperature, humidity
}
\description{
5 days weather forecast
}
\examples{
fiveDay("vancouver")
fiveDay("vancouver, ca")
fiveDay("Vancouver, CA")
fiveDay("VANCOUVER, CA")
}
