% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w3.R
\name{historical_weather}
\alias{historical_weather}
\title{Historical Weather Data}
\usage{
historical_weather(x)
}
\arguments{
\item{x}{DT, date from the 5 previous days (unix time, UTC time zone)}
}
\value{
ggplot
}
\description{
Historical Weather Data
}
\examples{
historical_weather("1613547787")
}
