% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w4.R
\name{search_weather}
\alias{search_weather}
\title{Search weather data}
\usage{
search_weather(x)
}
\arguments{
\item{x}{City}
}
\value{
City name, weather description, tempreture, humidity
}
\description{
Search weather data
}
\examples{
search_weather("london, UK")
search_weather("london")
}
